requirements.txt
progressbar2
opt_einsum
soundfile
librosa